+++ src/frontend/apps/e2e/__tests__/app-impress/doc-table-content.spec.ts
@@ -1,6 +1,6 @@
 import { expect, test } from '@playwright/test';
 
-import { createDoc, goToGridDoc, verifyDocName } from './common';
+import { createDoc, verifyDocName } from './common';
 
 test.beforeEach(async ({ page }) => {
   await page.goto('/');
@@ -19,20 +19,13 @@ test.describe('Doc Table Content', () => {
 
     await verifyDocName(page, randomDoc);
 
-    await page.getByLabel('Open the document options').click();
-    await page
-      .getByRole('button', {
-        name: 'Table of contents',
-      })
-      .click();
-
-    const panel = page.getByLabel('Document panel');
     const editor = page.locator('.ProseMirror');
 
     await editor.locator('.bn-block-outer').last().fill('/');
+
     await page.getByText('Heading 1').click();
-    await page.keyboard.type('Hello World');
-    await editor.getByText('Hello').dblclick();
+    await page.keyboard.type('Level 1');
+    await editor.getByText('Level 1').dblclick();
     await page.getByRole('button', { name: 'Strike' }).click();
 
     await page.locator('.bn-block-outer').first().click();
@@ -40,101 +33,44 @@ test.describe('Doc Table Content', () => {
     await page.locator('.bn-block-outer').last().click();
 
     // Create space to fill the viewport
-    for (let i = 0; i < 10; i++) {
+    for (let i = 0; i < 2; i++) {
       await page.keyboard.press('Enter');
     }
 
     await editor.locator('.bn-block-outer').last().fill('/');
     await page.getByText('Heading 2').click();
-    await page.keyboard.type('Super World', { delay: 100 });
+    await page.keyboard.type('Level 2');
 
     await page.locator('.bn-block-outer').last().click();
 
     // Create space to fill the viewport
-    for (let i = 0; i < 10; i++) {
+    for (let i = 0; i < 2; i++) {
       await page.keyboard.press('Enter');
     }
 
     await editor.locator('.bn-block-outer').last().fill('/');
     await page.getByText('Heading 3').click();
-    await page.keyboard.type('Another World');
+    await page.keyboard.type('Level 3');
 
-    const hello = panel.getByText('Hello World');
-    const superW = panel.getByText('Super World');
-    const another = panel.getByText('Another World');
+    expect(true).toBe(true);
 
-    await expect(hello).toBeVisible();
-    await expect(hello).toHaveCSS('font-size', /17/);
-    await expect(hello).toHaveAttribute('aria-selected', 'true');
+    const summaryContainer = page.locator('#summaryContainer');
+    await summaryContainer.hover();
 
-    await expect(superW).toBeVisible();
-    await expect(superW).toHaveCSS('font-size', /14/);
-    await expect(superW).toHaveAttribute('aria-selected', 'false');
+    const level1 = summaryContainer.getByText('Level 1');
+    const level2 = summaryContainer.getByText('Level 2');
+    const level3 = summaryContainer.getByText('Level 3');
 
-    await expect(another).toBeVisible();
-    await expect(another).toHaveCSS('font-size', /12/);
-    await expect(another).toHaveAttribute('aria-selected', 'false');
+    await expect(level1).toBeVisible();
+    await expect(level1).toHaveCSS('padding', /4px 0px/);
+    await expect(level1).toHaveAttribute('aria-selected', 'true');
 
-    await hello.click();
+    await expect(level2).toBeVisible();
+    await expect(level2).toHaveCSS('padding-left', /14.4px/);
+    await expect(level2).toHaveAttribute('aria-selected', 'false');
 
-    await expect(editor.getByText('Hello World')).toBeInViewport();
-    await expect(hello).toHaveAttribute('aria-selected', 'true');
-    await expect(superW).toHaveAttribute('aria-selected', 'false');
-
-    await another.click();
-
-    await expect(editor.getByText('Hello World')).not.toBeInViewport();
-    await expect(hello).toHaveAttribute('aria-selected', 'false');
-    await expect(superW).toHaveAttribute('aria-selected', 'true');
-
-    await panel.getByText('Back to top').click();
-    await expect(editor.getByText('Hello World')).toBeInViewport();
-    await expect(hello).toHaveAttribute('aria-selected', 'true');
-    await expect(superW).toHaveAttribute('aria-selected', 'false');
-
-    await panel.getByText('Go to bottom').click();
-    await expect(editor.getByText('Hello World')).not.toBeInViewport();
-    await expect(superW).toHaveAttribute('aria-selected', 'true');
-  });
-
-  test('it checks that table contents panel is opened automaticaly if more that 2 headings', async ({
-    page,
-    browserName,
-  }) => {
-    const [randomDoc] = await createDoc(
-      page,
-      'doc-table-content',
-      browserName,
-      1,
-    );
-
-    await verifyDocName(page, randomDoc);
-    await expect(page.getByLabel('Open the panel')).toBeHidden();
-
-    const editor = page.locator('.ProseMirror');
-
-    await editor.locator('.bn-block-outer').last().fill('/');
-    await page.getByText('Heading 1').click();
-    await page.keyboard.type('Hello World', { delay: 100 });
-
-    await page.keyboard.press('Enter');
-
-    await editor.locator('.bn-block-outer').last().fill('/');
-    await page.getByText('Heading 2').click();
-    await page.keyboard.type('Super World', { delay: 100 });
-
-    await goToGridDoc(page, {
-      title: randomDoc,
-    });
-
-    await expect(page.getByLabel('Close the panel')).toBeVisible();
-
-    const panel = page.getByLabel('Document panel');
-    await expect(panel.getByText('Hello World')).toBeVisible();
-    await expect(panel.getByText('Super World')).toBeVisible();
-
-    await page.getByLabel('Close the panel').click();
-
-    await expect(panel).toHaveAttribute('aria-hidden', 'true');
+    await expect(level3).toBeVisible();
+    await expect(level3).toHaveCSS('padding-left', /24px/);
+    await expect(level3).toHaveAttribute('aria-selected', 'false');
   });
 });