+++ src/frontend/apps/impress/src/components/separators/HorizontalSeparator.tsx
@@ -9,10 +9,12 @@ export enum SeparatorVariant {
 
 type Props = {
   variant?: SeparatorVariant;
+  $withPadding?: boolean;
 };
 
 export const HorizontalSeparator = ({
   variant = SeparatorVariant.LIGHT,
+  $withPadding = true,
 }: Props) => {
   const { colorsTokens } = useCunninghamTheme();
 
@@ -20,7 +22,7 @@ export const HorizontalSeparator = ({
     <Box
       $height="1px"
       $width="100%"
-      $margin={{ vertical: 'base' }}
+      $margin={{ vertical: $withPadding ? 'base' : 'none' }}
       $background={
         variant === SeparatorVariant.DARK
           ? '#e5e5e533'