+++ src/frontend/apps/impress/src/features/docs/doc-editor/components/DocEditor.tsx
@@ -2,9 +2,10 @@ import { Alert, Loader, VariantType } from '@openfun/cunningham-react';
 import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
+import { css } from 'styled-components';
 import * as Y from 'yjs';
 
-import { Box, Card, Text, TextErrors } from '@/components';
+import { Box, Text, TextErrors } from '@/components';
 import { useCunninghamTheme } from '@/cunningham';
 import { DocHeader, DocVersionHeader } from '@/features/docs/doc-header/';
 import {
@@ -12,11 +13,11 @@ import {
   base64ToBlocknoteXmlFragment,
   useDocStore,
 } from '@/features/docs/doc-management';
+import { TableContent } from '@/features/docs/doc-table-content/';
 import { Versions, useDocVersion } from '@/features/docs/doc-versioning/';
 import { useResponsiveStore } from '@/stores';
 
 import { BlockNoteEditor, BlockNoteEditorVersion } from './BlockNoteEditor';
-import { IconOpenPanelEditor, PanelEditor } from './PanelEditor';
 
 interface DocEditorProps {
   doc: Doc;
@@ -25,9 +26,9 @@ interface DocEditorProps {
 
 export const DocEditor = ({ doc, versionId }: DocEditorProps) => {
   const { t } = useTranslation();
-  const { isMobile } = useResponsiveStore();
+  const { isDesktop } = useResponsiveStore();
 
-  const isVersion = versionId && typeof versionId === 'string';
+  const isVersion = !!versionId && typeof versionId === 'string';
 
   const { colorsTokens } = useCunninghamTheme();
 
@@ -40,41 +41,49 @@ export const DocEditor = ({ doc, versionId }: DocEditorProps) => {
 
   return (
     <>
-      {isVersion ? (
-        <DocVersionHeader title={doc.title} />
-      ) : (
-        <DocHeader doc={doc} />
-      )}
-
-      {!doc.abilities.partial_update && (
-        <Box $width="100%" $margin={{ all: 'small', top: 'none' }}>
-          <Alert type={VariantType.WARNING}>
-            {t(`Read only, you cannot edit this document.`)}
-          </Alert>
+      {isDesktop && !isVersion && (
+        <Box
+          $position="absolute"
+          $css={css`
+            top: 72px;
+            right: 20px;
+          `}
+        >
+          <TableContent />
         </Box>
       )}
+      <Box $maxWidth="868px" $width="100%">
+        <Box $padding={{ horizontal: '54px' }}>
+          {isVersion ? (
+            <DocVersionHeader title={doc.title} />
+          ) : (
+            <DocHeader doc={doc} />
+          )}
+        </Box>
 
-      <Box
-        $background={colorsTokens()['primary-bg']}
-        $direction="row"
-        $width="100%"
-        $css="overflow-x: clip; flex: 1;"
-        $position="relative"
-      >
-        <Card
-          $padding={isMobile ? 'small' : 'big'}
-          $css="flex:1;"
-          $overflow="auto"
+        {!doc.abilities.partial_update && (
+          <Box $width="100%" $margin={{ all: 'small', top: 'none' }}>
+            <Alert type={VariantType.WARNING}>
+              {t(`Read only, you cannot edit this document.`)}
+            </Alert>
+          </Box>
+        )}
+
+        <Box
+          $background={colorsTokens()['primary-bg']}
+          $direction="row"
+          $width="100%"
+          $css="overflow-x: clip; flex: 1;"
           $position="relative"
         >
-          {isVersion ? (
-            <DocVersionEditor docId={doc.id} versionId={versionId} />
-          ) : (
-            <BlockNoteEditor doc={doc} provider={provider} />
-          )}
-          {!isMobile && !isVersion && <IconOpenPanelEditor />}
-        </Card>
-        {!isVersion && <PanelEditor />}
+          <Box $css="flex:1;" $overflow="auto" $position="relative">
+            {isVersion ? (
+              <DocVersionEditor docId={doc.id} versionId={versionId} />
+            ) : (
+              <BlockNoteEditor doc={doc} provider={provider} />
+            )}
+          </Box>
+        </Box>
       </Box>
     </>
   );