+++ src/frontend/apps/impress/src/features/docs/doc-table-content/components/TableContent.tsx
@@ -1,21 +1,22 @@
 import { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
+import { css } from 'styled-components';
 
-import { Box, BoxButton, Text } from '@/components';
+import { Box, Icon, Text } from '@/components';
 import { useEditorStore, useHeadingStore } from '@/features/docs/doc-editor';
 import { MAIN_LAYOUT_ID } from '@/layouts/conf';
-import { useResponsiveStore } from '@/stores';
 
 import { Heading } from './Heading';
 
 export const TableContent = () => {
   const { headings } = useHeadingStore();
   const { editor } = useEditorStore();
-  const { isMobile } = useResponsiveStore();
-  const { t } = useTranslation();
+
   const [headingIdHighlight, setHeadingIdHighlight] = useState<string>();
 
-  // To highlight the first heading in the viewport
+  const { t } = useTranslation();
+  const [isHover, setIsHover] = useState(false);
+
   useEffect(() => {
     const handleScroll = () => {
       if (!headings) {
@@ -62,69 +63,84 @@ export const TableContent = () => {
   }
 
   return (
-    <Box $padding={{ all: 'small', right: 'none' }} $maxHeight="95%">
-      <Box $overflow="auto" $padding={{ left: '2px' }}>
-        {headings?.map(
-          (heading) =>
-            heading.contentText && (
-              <Heading
-                editor={editor}
-                headingId={heading.id}
-                level={heading.props.level}
-                text={heading.contentText}
-                key={heading.id}
-                isHighlight={headingIdHighlight === heading.id}
-              />
-            ),
-        )}
-      </Box>
-      <Box
-        $height="1px"
-        $width="auto"
-        $background="#e5e5e5"
-        $margin={{ vertical: 'small' }}
-        $css="flex: none;"
-      />
-      <BoxButton
-        onClick={() => {
-          // With mobile the focus open the keyboard and the scroll is not working
-          if (!isMobile) {
-            editor.focus();
-          }
+    <Box
+      onMouseEnter={() => {
+        setIsHover(true);
+        setTimeout(() => {
+          const element = document.getElementById(
+            `heading-${headingIdHighlight}`,
+          );
 
-          document.querySelector(`.bn-editor`)?.scrollIntoView({
+          element?.scrollIntoView({
             behavior: 'smooth',
-            block: 'start',
+            inline: 'center',
+            block: 'center',
           });
-        }}
-        $align="start"
-      >
-        <Text $theme="primary" $padding={{ vertical: 'xtiny' }}>
-          {t('Back to top')}
-        </Text>
-      </BoxButton>
-      <BoxButton
-        onClick={() => {
-          // With mobile the focus open the keyboard and the scroll is not working
-          if (!isMobile) {
-            editor.focus();
-          }
+        }, 250); // 300ms is the transition time of the box
+      }}
+      onMouseLeave={() => {
+        setIsHover(false);
+      }}
+      id="summaryContainer"
+      $effect="show"
+      $width="40px"
+      $height="40px"
+      $zIndex={1000}
+      $align="center"
+      $padding="xs"
+      $justify="center"
+      $css={css`
+        border: 1px solid #ccc;
+        overflow: hidden;
+        border-radius: var(--c--theme--spacings--3xs);
+        background: var(--c--theme--colors--greyscale-000);
 
-          document
-            .querySelector(
-              `.bn-editor > .bn-block-group > .bn-block-outer:last-child`,
-            )
-            ?.scrollIntoView({
-              behavior: 'smooth',
-              block: 'start',
-            });
-        }}
-        $align="start"
-      >
-        <Text $theme="primary" $padding={{ vertical: 'xtiny' }}>
-          {t('Go to bottom')}
-        </Text>
-      </BoxButton>
+        &:hover {
+          overflow-y: auto;
+          display: flex;
+          flex-direction: column;
+          justify-content: flex-start;
+          align-items: flex-start;
+          gap: var(--c--theme--spacings--2xs);
+          width: 200px;
+          height: auto;
+          max-height: calc(100vh - 60px - 15vh);
+        }
+      `}
+    >
+      {!isHover && (
+        <Box $justify="center" $align="center">
+          <Icon iconName="list" $theme="primary" $variation="800" />
+        </Box>
+      )}
+      {isHover && (
+        <Box $width="100%">
+          <Box
+            $margin={{ bottom: '20px' }}
+            $direction="row"
+            $justify="space-between"
+            $align="center"
+          >
+            <Text $weight="bold" $variation="800" $theme="primary">
+              {t('Summary')}
+            </Text>
+            <Icon iconName="list" $theme="primary" $variation="800" />
+          </Box>
+          {headings?.map(
+            (heading) =>
+              heading.contentText && (
+                <Heading
+                  editor={editor}
+                  headingId={heading.id}
+                  level={heading.props.level}
+                  text={heading.contentText}
+                  key={heading.id}
+                  isHighlight={headingIdHighlight === heading.id}
+                />
+              ),
+          )}
+        </Box>
+      )}
     </Box>
   );
 };